requirements.txt
# Core PDF Processing
pypdf==4.0.1
pdfplumber==0.11.0
PyMuPDF==1.23.26

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM Integration (Optional - for advanced extraction)
# LiteLLM provides unified interface for all LLM providers
litellm==1.50.0

# Specific LLM providers (install as needed)
google-generativeai==0.3.2  # For Google Gemini
openai==1.10.0              # For OpenAI GPT models
anthropic==0.34.0           # For Claude models

# Legacy LangChain support (optional)
langchain==0.1.4
langchain-openai==0.0.5

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
tqdm==4.66.1

# Logging
structlog==24.1.0

# Jupyter for analysis
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.1
